ADDED preferences to the application.
Preferences are in the res/xml file. Pretty handy to setup and modify, had a bit of trouble setting up Google Maps activity to load but it works, just got that rendering bug me and pete had, probably down to manifest. Tried a few things but no joy on rendering the map.

The actual override is set from the edit text preference, on the very first install its always the default value but after redeploying to the emulator it stays on the string that it was set and I have confirmed this through stack trace.
Using preferences manager  to set the override as String to what is saved in preferences on the DismissAlarmActivity did not work and crashes the app when the alarm rings during run time but sure we can sort it out tomorrow/later today. Set it back to hardcoded value so it should work for now!

COMMENTED most of my XML stuff and refactored to remove redundant location stuff, culled out a lot but still let the app run. Replaced the location alarm fragment with text stub, disabled add new location etc.

REFACTORED layout files so they appear in order "activity_etc, dialog_etc" neater and easier to read, we should do the same for class files!
ADDED @string for a lot of the text in my layouts as well as tweaking layout/look of the cards, think its a little better.